Updated phpdoc comments.
The only part which i still don't like are the __constructor, __get and
__set methods. These are new in PHP5, so probably these aren't really
supported yet. These methods are called by the PHP engine and not by the
user.

The user is calling $rc4->key = "blabla"; and behind the scenes PHP is
calling the __set method. Maybe I should introduce a virtual variable and
mark __get and __set are private?


git-svn-id: http://svn.php.net/repository/pear/packages/Crypt_Rc4/trunk@163501 c90b9560-bf6c-de11-be94-00142212c4b1